Create app.config.ts
