+++ types/Preset.ts
@@ -1,0 +1,7 @@
+export interface Preset {
+  label: string,
+  width: number,
+  widthRatio: number,
+  height: number,
+  heightRatio: number
+}
